Add /account-manager/account/v1/avatar
